LPS-33269 & LPS-33085 Fix compile error

+++ portal-impl/src/com/liferay/portal/upgrade/v6_2_0/UpgradePolls.java
@@ -170,7 +170,7 @@ public class UpgradePolls extends UpgradeProcess {
 					continue;
 				}
 
-				long groupId = questionArray[0];
+				long groupId = (Long)questionArray[0];
 
 				runSQL(
 					"update PollsVote set groupId = " + groupId +
